requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
langchain>=0.3.7
langchain-core>=0.3.15
langgraph>=0.1.20
langgraph-checkpoint-sqlite>=1.0.0
langchain-google-genai>=2.0.0
google-generativeai>=0.8.3
pydantic>=2.9.0
python-dotenv>=1.0.1
requests>=2.32.3
python-multipart>=0.0.9
httpx>=0.27.2
pytest>=8.3.2
pytest-asyncio>=0.23.8
aiosqlite>=0.21.0
